post mediImg API cameraScreen으로 옮기고, 로딩 스크린 다시 구현 is Loading state를 camerascreen에 만들어주고 전체 화면을 한번더 분기해줌
